task 104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-int x;
-int t;
-int pt;
-int ppt;
+unsigned long int x;
+unsigned long int t;
+unsigned long int pt;
+unsigned long int ppt;
 
 t = 0;
 pt = 1;
@@ -17,7 +17,7 @@ ppt = 0;
 for (x = 0; x <= 97; x++)
 {
 t = ppt + pt;
-printf("%d", t);
+printf("%lu", t);
 ppt = pt;
 pt = t;
 if (x != 97)
